Clean out non-cm files
Got rid of my Pearson stuff and some older platform prototypes. Trying
to get this repo to be only for the marketing site. Also removed CNAME
as it seemed to be causing problems, fighting for control from my
personal repo. We’ll fix that once this is the way we want it.